Worker branch (#27)
* Implement progress tracking feature with IProgressService and ProgressController

* Refactor progress tracking by replacing IProgressService with IUserProgressService and implementing UserProgressController for enhanced user progress management.

* Enhance UserProgressController with authorization and refactor UserProgressService for improved readability

* Replace [Authorize] attribute with [RequireSubscription] in UserProgressController for subscription-based access control

* temp
