+++ src/Quiz.CSharp.Api/Controllers/CollectionsController.cs
@@ -2,7 +2,7 @@ namespace Quiz.CSharp.Api.Controllers;
 
 using Microsoft.AspNetCore.Mvc;
 using Quiz.CSharp.Api.Contracts;
-using Quiz.CSharp.Api.Services;
+using Quiz.CSharp.Api.Services.Abstractions;
 using Quiz.Shared.Contracts;
 
 [ApiController]
