+++ src/Quiz.CSharp.Api/Services/Abstractions/IAnswerService.cs
@@ -1,4 +1,4 @@
-namespace Quiz.CSharp.Api.Services;
+namespace Quiz.CSharp.Api.Services.Abstractions;
 
 using Quiz.CSharp.Api.Contracts;
 using Quiz.Shared.Common;
@@ -10,5 +10,5 @@ public interface IAnswerService
         string answer,
         int timeSpentSeconds,
         CancellationToken cancellationToken = default);
-    Task<UserAnswerResponse?> GetLatestAnswerAsync(int questionId, CancellationToken cancellationToken = default);
+    Task<UserAnswerResponse?> GetLatestAnswerOrDefaultAsync(int questionId, CancellationToken cancellationToken = default);
 }
