+++ src/Quiz.CSharp.Api/Services/Abstractions/IQuestionService.cs
@@ -1,4 +1,4 @@
-namespace Quiz.CSharp.Api.Services;
+namespace Quiz.CSharp.Api.Services.Abstractions;
 
 using Quiz.CSharp.Api.Contracts;
 using Quiz.Shared.Common;
