--- src/Quiz.CSharp.Api/Services/IAnswerValidator.cs
@@ -1,8 +1,0 @@
-namespace Quiz.CSharp.Api.Services;
-
-using Quiz.CSharp.Data.Entities;
-
-public interface IAnswerValidator
-{
-    Task<bool> ValidateAnswerAsync(Question question, string userAnswer, CancellationToken cancellationToken = default);
-}
