--- src/Quiz.CSharp.Api/Services/ICollectionService.cs
@@ -1,8 +1,0 @@
-namespace Quiz.CSharp.Api.Services;
-
-using Quiz.CSharp.Api.Contracts;
-
-public interface ICollectionService
-{
-    Task<List<CollectionResponse>> GetCollectionsAsync(CancellationToken cancellationToken = default);
-}
