+++ src/Quiz.CSharp.Api/Services/QuestionService.cs
@@ -3,12 +3,14 @@ namespace Quiz.CSharp.Api.Services;
 using AutoMapper;
 using Quiz.CSharp.Api.Contracts;
 using Quiz.CSharp.Data.Entities;
-using Quiz.CSharp.Data.Services;
+using Quiz.CSharp.Data.Repositories.Abstractions;
 using Quiz.Shared.Authentication;
 using Quiz.Shared.Common;
+using Quiz.CSharp.Api.Services.Abstractions;
 
 public sealed class QuestionService(
-    ICSharpRepository repository,
+    IAnswerRepository answerRepository,
+    IQuestionRepository questionRepository,
     IMapper mapper,
     ICurrentUser currentUser) : IQuestionService
 {
@@ -18,7 +20,11 @@ public sealed class QuestionService(
         int pageSize,
         CancellationToken cancellationToken = default)
     {
-        var questions = await repository.GetQuestionsByCollectionAsync(collectionId, page, pageSize, cancellationToken);
+        var questions = await questionRepository.GetQuestionsByCollectionAsync(
+            collectionId,
+            page,
+            pageSize,
+            cancellationToken);
         var responses = new List<QuestionResponse>();
 
         foreach (var question in questions.Items)
@@ -27,7 +33,7 @@ public sealed class QuestionService(
 
             if (currentUser.IsAuthenticated && currentUser.UserId is not null)
             {
-                var previousAnswer = await repository.GetLatestAnswerAsync(
+                var previousAnswer = await answerRepository.GetLatestAnswerOrDefaultAsync(
                     currentUser.UserId,
                     question.Id,
                     cancellationToken);
@@ -49,12 +55,18 @@ public sealed class QuestionService(
             responses.Add(response);
         }
 
-        return new PaginatedResult<QuestionResponse>(responses, questions.TotalCount, questions.Page, questions.PageSize);
+        return new PaginatedResult<QuestionResponse>(
+            responses,
+            questions.TotalCount,
+            questions.Page,
+            questions.PageSize);
     }
 
-    public async Task<List<QuestionResponse>> GetPreviewQuestionsAsync(int collectionId, CancellationToken cancellationToken = default)
+    public async Task<List<QuestionResponse>> GetPreviewQuestionsAsync(
+        int collectionId,
+        CancellationToken cancellationToken = default)
     {
-        var questions = await repository.GetPreviewQuestionsAsync(collectionId, cancellationToken);
+        var questions = await questionRepository.GetPreviewQuestionsAsync(collectionId, cancellationToken);
         return mapper.Map<List<QuestionResponse>>(questions);
     }
 }
