+++ src/Quiz.CSharp.Data/Data/CSharpDbContext.cs
@@ -1,18 +1,8 @@
-namespace Quiz.CSharp.Data;
+namespace Quiz.CSharp.Data.Data;
 
 using Microsoft.EntityFrameworkCore;
 using Quiz.CSharp.Data.Entities;
 using System.Reflection;
-
-public interface ICSharpDbContext
-{
-    DbSet<Collection> Collections { get; set; }
-    DbSet<Question> Questions { get; set; }
-    DbSet<UserAnswer> UserAnswers { get; set; }
-    DbSet<UserProgress> UserProgress { get; set; }
-
-    Task<int> SaveChangesAsync(CancellationToken cancellationToken = default);
-}
 
 public sealed class CSharpDbContext(DbContextOptions<CSharpDbContext> options) : DbContext(options), ICSharpDbContext
 {
