+++ src/Quiz.CSharp.Data/Extensions/ServiceCollectionExtensions.cs
@@ -4,6 +4,9 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Quiz.CSharp.Data.Services;
+using Quiz.CSharp.Data.Repositories.Abstractions;
+using Quiz.CSharp.Data.Repositories;
+using Quiz.CSharp.Data.Data;
 
 public static class ServiceCollectionExtensions
 {
@@ -15,7 +18,10 @@ public static class ServiceCollectionExtensions
             options.UseNpgsql(configuration.GetConnectionString("CSharpQuiz"))
                    .UseSnakeCaseNamingConvention());
 
-        services.AddScoped<ICSharpRepository, CSharpRepository>();
+        services.AddScoped<IAnswerRepository, AnswerRepository>();
+        services.AddScoped<ICollectionRepository, CollectionRepository>();
+        services.AddScoped<IQuestionRepository, QuestionRepository>();
+        services.AddScoped<IUserProgressRepository, UserProgressRepository>();
         services.AddHostedService<DatabaseMigrationService>();
         services.AddHostedService<DataSeedingService>();
