+++ src/Quiz.CSharp.Data/Services/DatabaseMigrationService.cs
@@ -4,6 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using Quiz.CSharp.Data.Data;
 
 public class DatabaseMigrationService(
         IServiceProvider serviceProvider,
